compiler: Implement indentation-aware block parsing
This required fixing a lexer issue with parsing quoted strings (columns have
to be correct to properly handle indentation).

Block parsing is currently rather strict - it requires you to maintain the
indentation precisely within a block, and block has to be indented relative
to the "starting" line. All blocks that don't start on their own line contain
a single expression.