compiler: Reach fixpoint when rewrites stop happening
Previously we could introduce constraints between fresh unknown types that
would never be satisfied; while it would be great to stop doing that, for
now it's not possible - so propagate pass returns the number of successfully
rewritten types instead.

This should prevent all cases where we never reach a fixpoint.