runtime: Implement native coroutine support
This is 100x+ faster than setjmp (which in itself is 1.3x faster than ucontext).

The speed benefit is mostly due to not saving signal mask. A downside is that
we'll have to implement context switching per-architecture (or rather per-ABI),
however the performance benefit is well worth it.

The change also slightly refactors scheduler implementation.