compiler: Add missing location information
Now most nodes have location data.

Note that for some (most?) cases the location is inaccurate - ideally we should
use spans that cover the entire node, but for most nodes we only use the first
token.

This is not important for debug info but is very important for error messages,
so it needs to be fixed at some point. Right now it's a bit cumbersome to do
this because it's hard to get a location from Ast node without knowing its type.

This requires extending union mechanism to allow for shared data, which is
probably a good idea anyway...