runtime: Switch to manual stack unwinding
_Unwind functions don't work well from a signal handler - the stack frames can
stop at the signal trampoline for some reason.

It's likely that they would work if we could initialize the unwind context with
data from signal context, but there is no interface for that. There is in
libunwind but it's big and scary.

For now let's switch to manual stack unwinding. It requires disabled frame
pointer omission and knowledge of stack bounds, but it's fast and simple.