compiler: Unify function typing rules
Now any function declaration can omit argument and return types and they are
inferred. In absence of a function body (for builtin/extern functions),
inference is impossible so the function type has to be known.

To keep the rules of typing the same between function types and declarations,
types have to specify the void return type explicitly now.