compiler: Implement simple function overloading
We currently only reduce the candidate list for calls since otherwise we'd have
to track multiple targets throughout typeck which is too complicated for now.

The overloading will sometimes not work correctly with inference if constraint
solver gives up, and instantiation needs to be reimplemented to be compatible
with the new flow.