compiler: Fix overload resolution with unknowns
Type constraint solver used to only match an unknown to one type and refuse
additional matches even if the matches are redundant; it now ignores redundant
matches.

Additionally, reducing the overload set results in possibility of different
type resolution in the future so we need to rerun the typeck to reach a true
fixed point.