compiler: Implement basic module importing
Restructure the compiler to have a queue of pending modules and two-phase
compilation - first all modules get parsed into an Ast, then all modules
are compiled.

This requires a global ordering of modules based on their imports since
we have to typecheck the dependencies before building the module - type
inference does not propagate type information to dependent modules. An
actual sorting implementation will follow; for now we use a hack.

Also restructure prelude handling so that driver does not know about it -
it's just a regular module.