compiler: Preserve argument types in Call node
Unfortunately, to generate a call to vararg function we also need the types of
all arguments. This information is not available in the callee type.

Even more unfortunately, this information can't be easily extracted from AST
nodes post-typeck so here's another hack to get that in codegen.

This will go away soon.