compiler: Rewrite typeck to maintain and use Ast::type
Right now this results in some assumptions about what other stages are doing
(e.g. if parser assigns an invalid type to the literal boolean typeck won't
detect that).

Ideally we'd fill everything that parser isn't sure about with Unknown types
and typeck would unify everything - it's not currently clear how expensive that
would be so use a conservative approach for now.

Various Ast hacks are still in but they're way less expensive and are now easy
to remove.