compiler: Support recursive types in codegenTypeInfo
This requires us to spell out the full type of struct type info since the
global variable has to be created with a complete type...

The only other way seems to require storing type info as global pointers which
would require 2 global variables per type info (pointer & data) or an init-time
initialization.

The current approach is more efficient and is not too cumbersome so we'll stick
with it for now.