compiler: Fix mangling for types that contain generics
Instead of forcing the caller to fully instantiate the type before mangling
which is error-prone, force the caller to pass an instantiation function.

This also means we don't have to explicitly instantiate types for functions
which saves a bit of time. Since getGenericInstance is single-pass we still
need to instantiate type args though.